Compile rlpvalue statically, defer install prefix
Small, non-frequently updated libs are best linked statically to remove
the brittle-ness of having the application dependent on dynamic
libraries scattered on the filesystem for reliability.

We defer the install-prefix to installation time due to rlpvalue
configuring a sub-project, univalue which does *not* inherit the prefix
value causing univalue to install into the system-wide prefix directory.

Additionally, call ./autogen.sh and ./configure explicitly with bash to
allow the build to be run on Windows from a MSYS-based shell reliably.