Add `callReadFunctionJSON` to return the JSON object from a RPC call
Allow returning of the original JSON object which allows the caller
avoid a memory allocation and compute cost of automatically
converting the JSON to a string. It also the caller to preserve
the JSON structure for lookup. Additionally, the caller can get
references to the underlying data avoiding additional allocations.