Overhaul APIs to be binary-centric first, then hex
Previous iteration of the API favoured returning hex from the interface
and working with a bunch of hex strings. What this leads to is

- Extra unneeded computation. Pretty much all libraries work with
binary data which meant a round-trip hex conversion everytime we produced data.

- String validation code is present everywhere. We have to ensure that
all the characters are hex and if they start with '0x' or even '0X' we
trim those out. This is a lot of extra work that is avoided by working
in binary.

I've relegated most of the dynamic containers to those that need
dynanism (like serializing a transaction to RLP since we have variable
data payloads). Everything else, especially hashes and cryptography are
now represented with fixed byte buffers.

All APIs generate binary data first, _then_ there're helper functions
that convert the binary to hex if hex is needed (e.g. serialising to JSON).